+++ packages/antd-management-fast-charts/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 2.6.162
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.7.55
+
+## 2.6.161
+
+### Patch Changes
+
+- Updated dependencies
+  - antd-management-fast-component@2.7.54
+
 ## 2.6.160
 
 ### Patch Changes