+++ packages/antd-management-fast-component/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 2.7.55
+
+### Patch Changes
+
+- update
+
+## 2.7.54
+
+### Patch Changes
+
+- update
+
 ## 2.7.53
 
 ### Patch Changes